refactor: refactor evaluation files
Remove a useless evaluation plugin file and rename a console script for better naming
